test: extract duplicated hypothesis strategies into a shared submodule
Moves shared test strategies (cells, elements, random_element_structures)
from tests/filter/test_aspect.py, tests/filter/test_volume.py,
tests/filter/test_distance.py, and tests/test_perturbations.py into
a centralized `tests/strategies` submodule that is accessible by pytest
but not shipped with the main package. Update references to use these new
shared strategies.
