+++ tests/filter/test_aspect.py
@@ -3,13 +3,7 @@ from ase.cell import Cell
 from assyst.filters import AspectFilter
 from hypothesis import given, strategies as st
 from pyxtal.lattice import generate_cellpara
-
-
-@st.composite
-def cells(draw):
-    ltype = st.sampled_from(["monoclinic", "triclinic", "orthorhombic", "tetragonal", "hexagonal", "trigonal", "cubic"])
-    volume = st.floats(min_value=5, max_value=100, allow_nan=False, allow_infinity=False)
-    return Cell.fromcellpar(generate_cellpara(draw(ltype), draw(volume)))
+from tests.strategies.strategies import cells
 
 
 @given(cells(), st.floats(0, exclude_min=True))
